+++ src/tailor/app.py
@@ -1185,13 +1185,16 @@ class Application(QtCore.QObject):
 
 def main():
     """Main entry point."""
-    qapp = QtWidgets.QApplication(sys.argv)
+    qapp = QtWidgets.QApplication()
     app = Application()
     app.ui.show()
     # Preflight
-    if not app.check_for_updates(silent=True):
-        # user does not want to install update so run the app
-        sys.exit(qapp.exec())
+    if not "--no-update-check" in sys.argv:
+        # will check for updates
+        if app.check_for_updates(silent=True):
+            # user wants to install available updates, so quit
+            return
+    sys.exit(qapp.exec())
 
 
 if __name__ == "__main__":